pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "fiddler-auditor"
version = "0.0.1.rc2"
authors = [
  { name="Fiddler Labs", email="support@fiddler.ai" },
]
description = "Auditing large language models made easy."
readme = "README.md"
requires-python = ">=3.7.12,<4.0"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "checklist==0.0.11",
    "pandas >=1.3.5,<2.0",
    "spacy-transformers >=1.1.8",
    "jinja2==3.1.2",
    "langchain>=0.0.158",
    "openai>=0.27.0",
    "sentence-transformers>=2.2.2",
]

[project.optional-dependencies]
test =[
  "pytest",
]

[tool.setuptools.package-data]
"modelauditor.assets.data" = ["*.csv", "*.txt"]
"modelauditor.reporting.templates" = ["report_template.html"]

[tool.setuptools.packages.find]
include = ["auditor*"]
